Extract ExchangeDataSource into a standalone class
