+++ core/trino-main/src/main/java/io/trino/exchange/ExchangeDataSource.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.exchange;
+
+import com.google.common.util.concurrent.ListenableFuture;
+import io.airlift.slice.Slice;
+import io.trino.operator.OperatorInfo;
+
+import java.io.Closeable;
+
+public interface ExchangeDataSource
+        extends Closeable
+{
+    Slice pollPage();
+
+    boolean isFinished();
+
+    ListenableFuture<Void> isBlocked();
+
+    void addInput(ExchangeInput input);
+
+    void noMoreInputs();
+
+    OperatorInfo getInfo();
+
+    @Override
+    void close();
+}
