+++ core/trino-main/src/main/java/io/trino/operator/ExchangeOperator.java
@@ -13,52 +13,32 @@
  */
 package io.trino.operator;
 
-import com.google.common.collect.ImmutableList;
 import com.google.common.util.concurrent.ListenableFuture;
-import com.google.common.util.concurrent.SettableFuture;
-import io.airlift.log.Logger;
 import io.airlift.slice.Slice;
 import io.trino.connector.CatalogHandle;
-import io.trino.exchange.DirectExchangeInput;
-import io.trino.exchange.ExchangeInput;
+import io.trino.exchange.ExchangeDataSource;
 import io.trino.exchange.ExchangeManagerRegistry;
-import io.trino.exchange.SpoolingExchangeInput;
-import io.trino.execution.TaskFailureListener;
-import io.trino.execution.TaskId;
+import io.trino.exchange.LazyExchangeDataSource;
 import io.trino.execution.buffer.PagesSerde;
 import io.trino.execution.buffer.PagesSerdeFactory;
 import io.trino.memory.context.LocalMemoryContext;
 import io.trino.metadata.Split;
 import io.trino.spi.Page;
 import io.trino.spi.connector.UpdatablePageSource;
-import io.trino.spi.exchange.ExchangeId;
-import io.trino.spi.exchange.ExchangeManager;
-import io.trino.spi.exchange.ExchangeSource;
-import io.trino.spi.exchange.ExchangeSourceHandle;
 import io.trino.split.RemoteSplit;
 import io.trino.sql.planner.plan.PlanNodeId;
 
-import java.io.Closeable;
-import java.net.URI;
-import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Supplier;
 
 import static com.google.common.base.Preconditions.checkArgument;
 import static com.google.common.base.Preconditions.checkState;
-import static com.google.common.util.concurrent.Futures.immediateVoidFuture;
-import static io.airlift.concurrent.MoreFutures.toListenableFuture;
 import static io.trino.connector.CatalogHandle.createRootCatalogHandle;
-import static java.lang.String.format;
 import static java.util.Objects.requireNonNull;
 
 public class ExchangeOperator
         implements SourceOperator
 {
-    private static final Logger log = Logger.get(ExchangeOperator.class);
-
     public static final CatalogHandle REMOTE_CATALOG_HANDLE = createRootCatalogHandle("$remote");
 
     public static class ExchangeOperatorFactory
@@ -235,340 +215,5 @@ public class ExchangeOperator
     public void close()
     {
         exchangeDataSource.close();
-    }
-
-    private interface ExchangeDataSource
-            extends Closeable
-    {
-        Slice pollPage();
-
-        boolean isFinished();
-
-        ListenableFuture<Void> isBlocked();
-
-        void addInput(ExchangeInput input);
-
-        void noMoreInputs();
-
-        OperatorInfo getInfo();
-
-        @Override
-        void close();
-    }
-
-    private static class LazyExchangeDataSource
-            implements ExchangeDataSource
-    {
-        private final TaskId taskId;
-        private final PlanNodeId sourceId;
-        private final DirectExchangeClientSupplier directExchangeClientSupplier;
-        private final LocalMemoryContext systemMemoryContext;
-        private final TaskFailureListener taskFailureListener;
-        private final RetryPolicy retryPolicy;
-        private final ExchangeManagerRegistry exchangeManagerRegistry;
-
-        private final SettableFuture<Void> initializationFuture = SettableFuture.create();
-        private final AtomicReference<ExchangeDataSource> delegate = new AtomicReference<>();
-        private final AtomicBoolean closed = new AtomicBoolean();
-
-        private LazyExchangeDataSource(
-                TaskId taskId,
-                PlanNodeId sourceId,
-                DirectExchangeClientSupplier directExchangeClientSupplier,
-                LocalMemoryContext systemMemoryContext,
-                TaskFailureListener taskFailureListener,
-                RetryPolicy retryPolicy,
-                ExchangeManagerRegistry exchangeManagerRegistry)
-        {
-            this.taskId = requireNonNull(taskId, "taskId is null");
-            this.sourceId = requireNonNull(sourceId, "sourceId is null");
-            this.directExchangeClientSupplier = requireNonNull(directExchangeClientSupplier, "directExchangeClientSupplier is null");
-            this.systemMemoryContext = requireNonNull(systemMemoryContext, "systemMemoryContext is null");
-            this.taskFailureListener = requireNonNull(taskFailureListener, "taskFailureListener is null");
-            this.retryPolicy = requireNonNull(retryPolicy, "retryPolicy is null");
-            this.exchangeManagerRegistry = requireNonNull(exchangeManagerRegistry, "exchangeManagerRegistry is null");
-        }
-
-        @Override
-        public Slice pollPage()
-        {
-            ExchangeDataSource dataSource = delegate.get();
-            if (dataSource == null) {
-                return null;
-            }
-            return dataSource.pollPage();
-        }
-
-        @Override
-        public boolean isFinished()
-        {
-            if (closed.get()) {
-                return true;
-            }
-            ExchangeDataSource dataSource = delegate.get();
-            if (dataSource == null) {
-                return false;
-            }
-            return dataSource.isFinished();
-        }
-
-        @Override
-        public ListenableFuture<Void> isBlocked()
-        {
-            if (closed.get()) {
-                return immediateVoidFuture();
-            }
-            if (!initializationFuture.isDone()) {
-                return initializationFuture;
-            }
-            ExchangeDataSource dataSource = delegate.get();
-            checkState(dataSource != null, "dataSource is expected to be initialized");
-            return dataSource.isBlocked();
-        }
-
-        @Override
-        public void addInput(ExchangeInput input)
-        {
-            boolean initialized = false;
-            synchronized (this) {
-                if (closed.get()) {
-                    return;
-                }
-                ExchangeDataSource dataSource = delegate.get();
-                if (dataSource == null) {
-                    if (input instanceof DirectExchangeInput) {
-                        DirectExchangeClient client = directExchangeClientSupplier.get(
-                                taskId.getQueryId(),
-                                new ExchangeId(format("direct-exchange-%s-%s", taskId.getStageId().getId(), sourceId)),
-                                systemMemoryContext,
-                                taskFailureListener,
-                                retryPolicy);
-                        dataSource = new DirectExchangeDataSource(client);
-                    }
-                    else if (input instanceof SpoolingExchangeInput) {
-                        SpoolingExchangeInput spoolingExchangeInput = (SpoolingExchangeInput) input;
-                        ExchangeManager exchangeManager = exchangeManagerRegistry.getExchangeManager();
-                        List<ExchangeSourceHandle> sourceHandles = spoolingExchangeInput.getExchangeSourceHandles();
-                        ExchangeSource exchangeSource = exchangeManager.createSource(sourceHandles);
-                        dataSource = new SpoolingExchangeDataSource(exchangeSource, sourceHandles, systemMemoryContext);
-                    }
-                    else {
-                        throw new IllegalArgumentException("Unexpected input: " + input);
-                    }
-                    delegate.set(dataSource);
-                    initialized = true;
-                }
-                dataSource.addInput(input);
-            }
-
-            if (initialized) {
-                initializationFuture.set(null);
-            }
-        }
-
-        @Override
-        public synchronized void noMoreInputs()
-        {
-            if (closed.get()) {
-                return;
-            }
-            ExchangeDataSource dataSource = delegate.get();
-            if (dataSource != null) {
-                dataSource.noMoreInputs();
-            }
-            else {
-                // to unblock when no splits are provided (and delegate hasn't been created)
-                close();
-            }
-        }
-
-        @Override
-        public OperatorInfo getInfo()
-        {
-            ExchangeDataSource dataSource = delegate.get();
-            if (dataSource == null) {
-                return null;
-            }
-            return dataSource.getInfo();
-        }
-
-        @Override
-        public void close()
-        {
-            synchronized (this) {
-                if (!closed.compareAndSet(false, true)) {
-                    return;
-                }
-                ExchangeDataSource dataSource = delegate.get();
-                if (dataSource != null) {
-                    dataSource.close();
-                }
-            }
-            initializationFuture.set(null);
-        }
-    }
-
-    private static class DirectExchangeDataSource
-            implements ExchangeDataSource
-    {
-        private final DirectExchangeClient directExchangeClient;
-
-        private DirectExchangeDataSource(DirectExchangeClient directExchangeClient)
-        {
-            this.directExchangeClient = requireNonNull(directExchangeClient, "directExchangeClient is null");
-        }
-
-        @Override
-        public Slice pollPage()
-        {
-            return directExchangeClient.pollPage();
-        }
-
-        @Override
-        public boolean isFinished()
-        {
-            return directExchangeClient.isFinished();
-        }
-
-        @Override
-        public ListenableFuture<Void> isBlocked()
-        {
-            return directExchangeClient.isBlocked();
-        }
-
-        @Override
-        public void addInput(ExchangeInput input)
-        {
-            DirectExchangeInput exchangeInput = (DirectExchangeInput) input;
-            directExchangeClient.addLocation(exchangeInput.getTaskId(), URI.create(exchangeInput.getLocation()));
-        }
-
-        @Override
-        public void noMoreInputs()
-        {
-            directExchangeClient.noMoreLocations();
-        }
-
-        @Override
-        public OperatorInfo getInfo()
-        {
-            return directExchangeClient.getStatus();
-        }
-
-        @Override
-        public void close()
-        {
-            directExchangeClient.close();
-        }
-    }
-
-    private static class SpoolingExchangeDataSource
-            implements ExchangeDataSource
-    {
-        // This field is not final to allow releasing the memory retained by the ExchangeSource instance.
-        // It is modified (assigned to null) when the ExchangeOperator is closed.
-        // It doesn't have to be declared as volatile as the nullification of this variable doesn't have to be immediately visible to other threads.
-        // However since close can be called at any moment this variable has to be accessed in a safe way (avoiding "check-then-use").
-        private ExchangeSource exchangeSource;
-        private final List<ExchangeSourceHandle> exchangeSourceHandles;
-        private final LocalMemoryContext systemMemoryContext;
-        private volatile boolean closed;
-
-        private SpoolingExchangeDataSource(
-                ExchangeSource exchangeSource,
-                List<ExchangeSourceHandle> exchangeSourceHandles,
-                LocalMemoryContext systemMemoryContext)
-        {
-            // this assignment is expected to be followed by an assignment of a final field to ensure safe publication
-            this.exchangeSource = requireNonNull(exchangeSource, "exchangeSource is null");
-            this.exchangeSourceHandles = ImmutableList.copyOf(requireNonNull(exchangeSourceHandles, "exchangeSourceHandles is null"));
-            this.systemMemoryContext = requireNonNull(systemMemoryContext, "systemMemoryContext is null");
-        }
-
-        @Override
-        public Slice pollPage()
-        {
-            ExchangeSource exchangeSource = this.exchangeSource;
-            if (exchangeSource == null) {
-                return null;
-            }
-
-            Slice data = exchangeSource.read();
-            systemMemoryContext.setBytes(exchangeSource.getMemoryUsage());
-
-            // If the data source has been closed in a meantime reset memory usage back to 0
-            if (closed) {
-                systemMemoryContext.setBytes(0);
-            }
-
-            return data;
-        }
-
-        @Override
-        public boolean isFinished()
-        {
-            ExchangeSource exchangeSource = this.exchangeSource;
-            if (exchangeSource == null) {
-                return true;
-            }
-            return exchangeSource.isFinished();
-        }
-
-        @Override
-        public ListenableFuture<Void> isBlocked()
-        {
-            ExchangeSource exchangeSource = this.exchangeSource;
-            if (exchangeSource == null) {
-                return immediateVoidFuture();
-            }
-            return toListenableFuture(exchangeSource.isBlocked());
-        }
-
-        @Override
-        public void addInput(ExchangeInput input)
-        {
-            SpoolingExchangeInput exchangeInput = (SpoolingExchangeInput) input;
-            // Only a single split is expected when external exchange is used.
-            // The engine adds the same split to every instance of the ExchangeOperator.
-            // Since the ExchangeDataSource is shared between ExchangeOperator instances
-            // the same split may be delivered multiple times.
-            checkState(
-                    exchangeInput.getExchangeSourceHandles().equals(exchangeSourceHandles),
-                    "split is expected to contain an identical exchangeSourceHandles list: %s != %s",
-                    exchangeInput.getExchangeSourceHandles(),
-                    exchangeSourceHandles);
-        }
-
-        @Override
-        public void noMoreInputs()
-        {
-            // Only a single split is expected when external exchange is used.
-            // Thus the assumption of "noMoreSplit" is made on construction.
-        }
-
-        @Override
-        public OperatorInfo getInfo()
-        {
-            return null;
-        }
-
-        @Override
-        public synchronized void close()
-        {
-            if (closed) {
-                return;
-            }
-            closed = true;
-            try {
-                exchangeSource.close();
-            }
-            catch (RuntimeException e) {
-                log.warn(e, "error closing exchange source");
-            }
-            finally {
-                exchangeSource = null;
-                systemMemoryContext.setBytes(0);
-            }
-        }
     }
 }
